Add option to generate launch scripts
Adds the option to generate batch and (posix) shell launch scripts for CLI applications to the MontiCore generator. For this purpose, two new command line flags are introduced: An input which expects the name of the CLI artifact and an output which specifies into which directory the scripts are generated.