Replace more occurrences of "http" with "https" links
Http URLs are generally considered insecure and prone to attacks. When a website serves a valid certificate, there is no reason to use a http link anymore. This commit replaces many occurrences of "http" in such links and removes a few remaining redundant copyright headers.

This commit does not change the links:
- http://mbse.se-rwth.de/ (because it serves mixed content),
- http://www.rumpe.eu/ (because it does not have a certificate), and
- links in license files. (Changing these shouldn't be a problem either, but I am not a lawyer, so I left them as-is for now.)